Do not calculate BV for camera objects
The camera rig is added as a child of the scene root object. Every time
the camera rig transform is updated, it results in the hierarchical bounding
volume being invalidated causing the path to the scene root
to be invalidated. Essentially scene graph is forced to recalculate the
bounding volume every time the camera rig is updated.

This change allows the camera rig to update itself without forcing the
bounding volume to be updated.